fix: document index tests depending on order
