+++ tests/connectors/retrievers/test_document_index_retriever.py
@@ -9,7 +9,5 @@ from intelligence_layer.connectors.retrievers.document_index_retriever import (
 def test_document_index_retriever(
     document_index_retriever: DocumentIndexRetriever,
 ) -> None:
-    documents = document_index_retriever.get_relevant_documents_with_scores(
-        "Who took part in the war?"
-    )
-    assert len(documents) == 2
+    documents = document_index_retriever.get_relevant_documents_with_scores("Coca-Cola")
+    assert len(documents) > 0
